Move socket handling to sockets.js and exclude app.js from testing
