--- src/test/server/app_test.js
@@ -1,4 +1,0 @@
-/**
- * Created by jacob on 20.04.17.
- */
-let app = require('../../server/app')
